Shit auf den neuesten Stand gebracht

+++ Fischer/Pflichtuebung1/Aufgabe2.c
@@ -61,7 +61,7 @@ float* sinusSignal(unsigned int N, unsigned int f, float a, unsigned int r) {
 }
 
 
-#define bufferlength 20
+#define bufferlength (int)20
 #define bufferkoeffizient ((float)1/(float)20)
